feat(components): new adBlock redirect component
affects: @haaretz/haaretz.co.il, @haaretz/htz-components

+++ packages/components/htz-components/src/components/Ads/DfpInjector.js
@@ -154,7 +154,10 @@ class DfpInjector extends Component {
   componentWillUnmount() {
     console.log('Destroy Slots');
     // eslint-disable-next-line no-undef
-    googletag.destroySlots();
+    if (googletag.destroySlots) {
+      // eslint-disable-next-line no-undef
+      googletag.destroySlots();
+    }
     // this.setState()
   }
